pyproject.toml
[tool.poetry]
name = "eburger"
version = "0.1.0"
description = ""
authors = ["tomie <admin@forefy.com>"]
readme = "README.md"
include = ["eburger/templates/*.yaml"]

[tool.poetry.scripts]
eburger = 'eburger.main:main'

[tool.poetry.dependencies]
python = "^3.11"
networkx = "^3.2.1"
pyvis = "^0.3.2"
argparse = "^1.4.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
